// src/pages/customer/master/data.d.ts
export interface TableListItem {
  expert_id?: number | string;
  expert_name?: string;
  identity_number?: string;
  mobile?: string;
  expert_type?: string;
  level_name?: string,
  job_title?: string;
  pro_title?: string;
  specialty?: string;
  avatar_url?: string ;
  answer_count: number | string;
  answers?: number | string;
  city_town?: number | string;
}